+++ eras/conway/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.21.0.0
 
+* Remove `asBoundedIntegralHKD`
+* Change lens type of `hkdCommitteeMinSizeL` and `ppCommitteeMinSize` from `Natural` to `Word16`
 * Add `HeaderProtVerTooHigh` predicate failure.
 * Change `STS` instance of `ConwayUTOXS`: use `PParams` as `Environment`
 * Remove `TotalDeposits` and `TxUTxODiff` data constructors from `ConwayUtxosEvent`